Fixed small bug in serverinfo
When you did the serverinfo command in a DM, it would print a different error message to what was expected. Although this isn't a huge issue as it doesn't interfere with the command itself, I felt it didn't look clean enough so I just had to move the if statement up a few lines to fix it.