Better deleting
Will now iterate through older posts to find posts. Previously, if the messages were older than the last 100 the commands failed.

Ideally tho, there should be a separate command that allows you to prune from messageid to message id. For example:
.prune endmessageid startmessageid - this should fetch messages starting at endmessageid, then fetch 100 messages each until it encounters the startmessageid, effectively deleting all the messages in between.

but for now this is better than the current functionality